Set timestampsLoaded.
